+++ hooks/useVercelChat.ts
@@ -24,11 +24,7 @@ import { useDeleteTrailingMessages } from "./useDeleteTrailingMessages";
 import { getFileContents } from "@/lib/sandboxes/getFileContents";
 import getMimeFromPath from "@/lib/files/getMimeFromPath";
 import { getChatPath } from "@/lib/chat/getChatPath";
-import { updateChat } from "@/lib/chats/updateChat";
-import {
-  shouldPersistChatModel,
-  type PersistedChatModel,
-} from "@/lib/chats/shouldPersistChatModel";
+import { usePersistSelectedModel } from "./usePersistSelectedModel";
 
 interface UseVercelChatProps {
   id: string;
@@ -70,9 +66,6 @@ export function useVercelChat({
   const userId = userData?.account_id || userData?.id; // Use account_id if available, fallback to id
   const artistId = selectedArtist?.account_id;
   const messagesLengthRef = useRef<number>();
-  // The {chatId, model} we last wrote to chats.model_id, so we only PATCH
-  // when the selection actually changes for the active chat.
-  const lastPersistedModelRef = useRef<PersistedChatModel>(null);
   const { addOptimisticConversation } = useConversationsProvider();
   const { data: availableModels = [] } = useAvailableModels();
   const [input, setInput] = useState("");
@@ -87,6 +80,16 @@ export function useVercelChat({
     sessionId,
   });
   const { authenticated, getAccessToken } = usePrivy();
+
+  // Persists the picker's selected model to chats.model_id; awaited before
+  // send so the workflow bills the selected model (it reads chats.model_id
+  // at request time), not the default.
+  const persistSelectedModel = usePersistSelectedModel({
+    sessionId,
+    chatId: transportChatId,
+    model,
+    getAccessToken,
+  });
 
   // Load artist files for mentions (from Supabase)
   const { files: allArtistFiles = [] } = useArtistFilesForMentions();
@@ -283,30 +286,8 @@ export function useVercelChat({
       files: nonAudioAttachments.length > 0 ? nonAudioAttachments : undefined,
     };
 
-    // Persist the picker's model to chats.model_id before sending. The
-    // chat-workflow path bills the model it reads from chats.model_id at
-    // request time, so the write must land first — awaited here so a new
-    // chat's very first turn bills the selected model, not the default.
-    if (
-      sessionId &&
-      transportChatId &&
-      shouldPersistChatModel(lastPersistedModelRef.current, transportChatId, model)
-    ) {
-      try {
-        const accessToken = await getAccessToken();
-        if (accessToken) {
-          await updateChat({
-            accessToken,
-            sessionId,
-            chatId: transportChatId,
-            modelId: model,
-          });
-          lastPersistedModelRef.current = { chatId: transportChatId, model };
-        }
-      } catch (error) {
-        console.error("Failed to persist selected model before send:", error);
-      }
-    }
+    // Land the selected model in chats.model_id before the send fires.
+    await persistSelectedModel();
 
     sendMessage(payload, { body: chatRequestBody, headers });
     setInput("");
